E_WeJob_Repo/requirements.txt
# Django 
django
djangorestframework
django-phonenumber-field

django-filter

